@@ -46,18 +46,17 @@
     "@electron-forge/maker-squirrel": "^6.0.1",
     "@electron-forge/maker-zip": "^6.0.1",
     "@electron-forge/plugin-webpack": "^6.0.1",
-    "@johmun/vue-tags-input": "^1.5.10",
+    "@sipec/vue3-tags-input": "^3.0.4",
     "@vercel/webpack-asset-relocator-loader": "^1.7.3",
-    "@vue/test-utils": "^1.0.0-beta.29",
     "axios": "^0.22.0",
     "babel-loader": "^8.2.2",
+    "bufferutil": "^4.0.9",
     "codemirror": "^6.0.1",
     "codemirror-lang-schemioscript": "^0.1.1",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
     "electron": "21.2.3",
     "electron-icon-maker": "^0.0.5",
-    "esm": "^3.2.25",
     "expect": "^27.4.2",
     "file-loader": "^0.11.2",
     "gifenc": "^1.0.3",
@@ -72,15 +71,13 @@
     "shortid": "^2.2.16",
     "style-loader": "^3.3.1",
     "thememirror": "^2.0.1",
-    "tiptap": "^1.22.3",
-    "tiptap-extensions": "^1.22.3",
-    "vue": "^2.6.10",
+    "tiny-emitter": "^2.1.0",
+    "utf-8-validate": "^5.0.10",
+    "vue": "^3.5.13",
     "vue-color": "2.8.1",
-    "vue-gtag": "^1.16.1",
-    "vue-loader": "^15.9.8",
-    "vue-router": "^2.7.0",
-    "vue-template-compiler": "^2.6.10",
-    "vuex": "^3.5.1",
+    "vue-loader": "^17.4.2",
+    "vue-router": "^4.5.0",
+    "vuex": "^4.1.0",
     "webpack": "^5.64.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.5.0"
