@@ -31,23 +31,26 @@
                 </div>
             </div>
 
-            <panel v-for="panel in filteredItemPanels" :key="panel.id" v-if="panel.items.length > 0" :name="panel.name">
-                <div class="item-menu">
-                    <div v-for="item in panel.items"
-                        class="item-container"
-                        :title="item.name"
-                        @mouseleave="stopPreviewItem(item)"
-                        @mouseover="showPreviewItem(item)"
-                        @mousedown="onItemMouseDown($event, item)"
-                        @dragstart="preventEvent"
-                        @drag="preventEvent"
-                        >
+            <template v-for="panel in filteredItemPanels">
+                <panel :key="panel.id" v-if="panel.items.length > 0" :name="panel.name">
+                    <div class="item-menu">
+                        <div v-for="item in panel.items"
+                            class="item-container"
+                            :title="item.name"
+                            @mouseleave="stopPreviewItem(item)"
+                            @mouseover="showPreviewItem(item)"
+                            @mousedown="onItemMouseDown($event, item)"
+                            @dragstart="preventEvent"
+                            @drag="preventEvent"
+                            >
 
-                        <img v-if="item.iconUrl" :src="item.iconUrl" width="42px" height="32px"/>
-                        <svg v-else-if="item.iconSVG" width="42px" height="32px" v-html="item.iconSVG"></svg>
+                            <img v-if="item.iconUrl" :src="item.iconUrl" width="42px" height="32px"/>
+                            <svg v-else-if="item.iconSVG" width="42px" height="32px" v-html="item.iconSVG"></svg>
+                        </div>
                     </div>
-                </div>
-            </panel>
+                </panel>
+            </template>
+
 
             <panel v-for="panel in extraShapeGroups" :key="panel.id"  :name="panel.name">
                 <div class="item-menu">
@@ -83,17 +86,18 @@
                     <span class="btn btn-secondary" @click="editArtModalShown = true" title="Edit art icons"><i class="fas fa-pencil-alt"></i></span>
                 </div>
                 <div class="item-menu">
-                    <div class="item-container"
-                        v-for="art in artList"
-                        v-if="!searchKeyword || safeTextMatchKeyword(art.name)"
-                        @mouseover="showPreviewArt(art)"
-                        @mouseleave="stopPreviewArt(art)"
-                        @mousedown="onArtMouseDown($event, art)"
-                        @dragstart="preventEvent"
-                        @drag="preventEvent"
-                        >
-                        <img :src="art.url"/>
-                    </div>
+                    <template v-for="art in artList">
+                        <div class="item-container"
+                            v-if="!searchKeyword || safeTextMatchKeyword(art.name)"
+                            @mouseover="showPreviewArt(art)"
+                            @mouseleave="stopPreviewArt(art)"
+                            @mousedown="onArtMouseDown($event, art)"
+                            @dragstart="preventEvent"
+                            @drag="preventEvent"
+                            >
+                            <img :src="art.url"/>
+                        </div>
+                    </template>
                 </div>
             </panel>
 
@@ -102,44 +106,50 @@
                     <span class="btn btn-secondary" @click="showExternalTemplateModal"><i class="fa-solid fa-square-up-right"></i> External Template...</span>
                 </div>
                 <div class="item-menu">
-                    <div class="item-container item-container-template"
-                        v-for="template in templates"
-                        v-if="!searchKeyword || safeTextMatchKeyword(template.name)"
-                        @mouseover="showPreviewTemplate(template)"
-                        @mouseleave="stopPreviewTemplate(template)"
-                        @mousedown="onTemplateMouseDown($event, template)"
-                        @dragstart="preventEvent"
-                        @drag="preventEvent"
-                        >
-                        <img v-if="template.preview" :src="template.preview"/>
-                        <img v-else :src="`${assetsPath}/images/missing-preview.svg`"/>
-                    </div>
-                </div>
-            </panel>
-
-            <panel v-for="artPack in filteredArtPacks" :key="artPack.id" v-if="artPack.icons.length > 0" :name="artPack.name" :closable="true" @close="closeArtPack(artPack)">
-                <div class="art-pack">
-                    <div class="art-pack-author" v-if="artPack.author">
-                        Created by
-                        <a v-if="artPack.link" :href="artPack.link">{{artPack.author}}</a>
-                        <span v-else>{{artPack.author}}</span>
-                    </div>
-                    <div class="item-menu">
-                        <div class="item-container"
-                            v-for="icon in artPack.icons"
-                            :key="icon.id"
-                            v-if="!searchKeyword || safeTextMatchKeyword(icon.name) || safeTextMatchKeyword(icon.description)"
-                            @mouseover="showPreviewArt(icon)"
-                            @mouseleave="stopPreviewArt(icon)"
-                            @mousedown="onArtMouseDown($event, icon)"
+                    <template v-for="template in templates">
+                        <div class="item-container item-container-template"
+                            v-if="!searchKeyword || safeTextMatchKeyword(template.name)"
+                            @mouseover="showPreviewTemplate(template)"
+                            @mouseleave="stopPreviewTemplate(template)"
+                            @mousedown="onTemplateMouseDown($event, template)"
                             @dragstart="preventEvent"
                             @drag="preventEvent"
                             >
-                            <img :src="icon.url" :title="`${icon.name} ${icon.description}`"/>
+                            <img v-if="template.preview" :src="template.preview"/>
+                            <img v-else :src="`${assetsPath}/images/missing-preview.svg`"/>
                         </div>
-                    </div>
+                    </template>
+
                 </div>
             </panel>
+
+            <template v-for="artPack in filteredArtPacks">
+                <panel :key="artPack.id" v-if="artPack.icons.length > 0" :name="artPack.name" :closable="true" @close="closeArtPack(artPack)">
+                    <div class="art-pack">
+                        <div class="art-pack-author" v-if="artPack.author">
+                            Created by
+                            <a v-if="artPack.link" :href="artPack.link">{{artPack.author}}</a>
+                            <span v-else>{{artPack.author}}</span>
+                        </div>
+                        <div class="item-menu">
+                            <template v-for="icon in artPack.icons">
+                                <div class="item-container"
+                                    :key="icon.id"
+                                    v-if="!searchKeyword || safeTextMatchKeyword(icon.name) || safeTextMatchKeyword(icon.description)"
+                                    @mouseover="showPreviewArt(icon)"
+                                    @mouseleave="stopPreviewArt(icon)"
+                                    @mousedown="onArtMouseDown($event, icon)"
+                                    @dragstart="preventEvent"
+                                    @drag="preventEvent"
+                                    >
+                                    <img :src="icon.url" :title="`${icon.name} ${icon.description}`"/>
+                                </div>
+                            </template>
+                        </div>
+                    </div>
+                </panel>
+            </template>
+
 
             <div class="section">
                 <span class="btn btn-secondary btn-block" @click="extraShapesModal.shown = true">More shapes &#8230;</span>
