@@ -27,11 +27,6 @@ import htmlSanitize from '../../../htmlSanitize';
 import utils from '../../utils';
 import RichTextEditor from '../RichTextEditor.vue';
 import { Keys, identifyKeyPress } from '../../events';
-import { Editor, EditorContent } from 'tiptap';
-import {
-    Blockquote, CodeBlock, HardBreak, Heading, OrderedList, BulletList, ListItem,
-    TodoItem, TodoList, Bold, Code, Italic, Strike, Underline, History,
-} from 'tiptap-extensions';
 import EditorEventBus from './EditorEventBus';
 
 
@@ -49,7 +44,7 @@ export default {
         markupDisabled : {type: Boolean, default: false},
         mouseDownId    : {type: Number},
     },
-    components: {RichTextEditor, EditorContent},
+    components: {RichTextEditor },
 
     beforeMount() {
         document.addEventListener('keydown', this.onKeyDown);
@@ -96,24 +91,6 @@ export default {
                 }
             }
             return editorCssStyle;
-        },
-
-        createEditor(text) {
-            const editor = new Editor({
-                disablePasteRules: true,
-                disableInputRules: true,
-                extensions: [
-                    new Blockquote(), new CodeBlock(), new HardBreak(), new Heading({ levels: [1, 2, 3] }), new BulletList(), new OrderedList(), new ListItem(),
-                    new TodoItem(), new TodoList(), new Bold(), new Code(), new Italic(), new Strike(), new Underline(), new History(), ],
-
-                autoFocus: true,
-                content: '',
-                onUpdate: (event) => {
-                    this.$emit('updated', event.getHTML());
-                }
-            });
-            editor.setContent(this.text, true, {preserveWhitespace: true})
-            return editor;
         },
 
         onKeyDown(event) {
