@@ -12,37 +12,39 @@
             @changed="emitArgChange('expression', arguments[0])" />
 
         <table>
-            <tr>
-                <td>If true:</td>
-                <td>
-                    <Dropdown
-                        :options="conditionBranchOptions"
-                        :auto-focus-search="args.success !== 'send-event'"
-                        :inline="true"
-                        :borderless="false"
-                        @selected="onSuccessSelected">
-                        <i :class="toIcon(args.success)"/>
-                        <span> {{ args.success | toPrettyName }} </span>
-                    </Dropdown>
-                    <input ref="sendEventInputSuccess" v-if="args.success === 'send-event'" type="text" :value="args.successEvent" class="condition-function-send-event"/>
-                </td>
-            </tr>
-            <tr>
-                <td>Else:</td>
-                <td>
-                    <Dropdown
-                        :options="conditionBranchOptions"
-                        :auto-focus-search="args.fail !== 'send-event'"
-                        :inline="true"
-                        :borderless="false"
-                        @selected="onFailSelected">
-                        <i :class="toIcon(args.fail)"/>
-                        <span> {{ args.fail | toPrettyName }} </span>
-                    </Dropdown>
-                    <input ref="sendEventInputFail" v-if="args.fail === 'send-event'" type="text" :value="args.failEvent" class="condition-function-send-event"/>
+            <tbody>
+                <tr>
+                    <td>If true:</td>
+                    <td>
+                        <Dropdown
+                            :options="conditionBranchOptions"
+                            :auto-focus-search="args.success !== 'send-event'"
+                            :inline="true"
+                            :borderless="false"
+                            @selected="onSuccessSelected">
+                            <i :class="toIcon(args.success)"/>
+                            <span> {{ args.success | toPrettyName }} </span>
+                        </Dropdown>
+                        <input ref="sendEventInputSuccess" v-if="args.success === 'send-event'" type="text" :value="args.successEvent" class="condition-function-send-event"/>
+                    </td>
+                </tr>
+                <tr>
+                    <td>Else:</td>
+                    <td>
+                        <Dropdown
+                            :options="conditionBranchOptions"
+                            :auto-focus-search="args.fail !== 'send-event'"
+                            :inline="true"
+                            :borderless="false"
+                            @selected="onFailSelected">
+                            <i :class="toIcon(args.fail)"/>
+                            <span> {{ args.fail | toPrettyName }} </span>
+                        </Dropdown>
+                        <input ref="sendEventInputFail" v-if="args.fail === 'send-event'" type="text" :value="args.failEvent" class="condition-function-send-event"/>
 
-                </td>
-            </tr>
+                    </td>
+                </tr>
+            </tbody>
         </table>
     </div>
 </template>
