@@ -17,7 +17,7 @@ module.exports = {
     },
     resolve: {
         alias: {
-            vue: 'vue/dist/vue.min.js',
+            // vue: 'vue/dist/vue.min.js',
             '@codemirror/state': path.resolve(__dirname, 'node_modules/@codemirror/state'),
             '@codemirror': path.resolve(__dirname, 'node_modules/@codemirror/'),
         }
