@@ -15,7 +15,7 @@ module.exports = {
     },
     resolve: {
         alias: {
-            vue: 'vue/dist/vue.js',
+            // vue: 'vue/dist/vue.js',
             '@codemirror/state': path.resolve(__dirname, 'node_modules/@codemirror/state'),
             '@codemirror': path.resolve(__dirname, 'node_modules/@codemirror/'),
         }
@@ -44,7 +44,7 @@ module.exports = {
     plugins: [
         new VueLoaderPlugin(),
         new webpack.IgnorePlugin({
-            resourceRegExp: /canvas/, 
+            resourceRegExp: /canvas/,
             contextRegExp: /jsdom$/
         }),
         new webpack.DefinePlugin({
